+++ nextgisweb/webmap/amd/ngw-webmap/ui/SearchPanel/SearchPanel.js
@@ -5,7 +5,7 @@ define([
     "ngw-pyramid/dynamic-panel/DynamicPanel",
     "dijit/layout/BorderContainer",
     "dojo/Deferred",
-    "dojo/request/script",
+    "dojo/request/xhr",
     "openlayers/ol",
     "dojo/on",
     "dojo/dom-class",
@@ -14,7 +14,7 @@ define([
     "dojo/_base/array",
     "ngw-feature-layer/FeatureStore",
     // settings
-    "@nextgisweb/pyramid/settings!feature_layer",
+    "@nextgisweb/pyramid/settings!webmap",
     // templates
     "dojo/text!./SearchPanel.hbs",
     // css
@@ -26,7 +26,7 @@ define([
     DynamicPanel,
     BorderContainer,
     Deferred,
-    script,
+    xhr,
     ol,
     on,
     domClass,
@@ -34,7 +34,7 @@ define([
     lang,
     array,
     FeatureStore,
-    featureLayersettings,
+    webmapSettings,
     template
 ) {
     return declare([DynamicPanel, BorderContainer],{
@@ -168,19 +168,24 @@ define([
                     }
                 }, this);
 
-                if (featureLayersettings.search.nominatim) {
+                if (webmapSettings.nominatim_enabled) {
                     var ndeferred = new Deferred();
 
                     deferred.then(lang.hitch(this, function (limit) {
-                        var NOMINATIM_SEARCH_URL = "https://nominatim.openstreetmap.org/search/";
-                        var CALLBACK = "json_callback";
+                        var NOMINATIM_SEARCH_URL = "https://nominatim.openstreetmap.org/search";
 
-                        jsonpArgs = {
-                            jsonp: CALLBACK,
-                            query: {format: "json", limit: "30", q: criteria}
-                        };
+                        var query = {format: "json", limit: "30", q: criteria};
+                        if (webmapSettings.nominatim_extent) {
+                            var extent = this.display.config.extent;
+                            query.bounded = "1";
+                            query.viewbox = extent.join(",");
+                        }
 
-                        script.get(NOMINATIM_SEARCH_URL, jsonpArgs).then(lang.hitch(this, function (data) {
+                        xhr.get(NOMINATIM_SEARCH_URL, {
+                            handleAs: "json",
+                            query: query,
+                            headers: { "X-Requested-With": null }
+                        }).then(lang.hitch(this, function (data) {
                             array.forEach(data, function (place) {
                                 if (limit > 0) {
                                     // Coordinates come in WGS84