+++ CHANGES.rst
@@ -9,6 +9,8 @@ Changes
 - Better support of KML: LIBKML GDAL driver is used when available.
 - Filtering features by ID in feature REST API.
 - Automatic downsampling of social preview image to 1600x630 pixels.
+- Disable (and enable) Nominatim search via settings in the control panel.
+- Ability to constraint Nominatim search area by web map initial extent.
 
 3.8.0
 -----